+++ .gitignore
@@ -36,7 +36,7 @@ activate/
 ENV/
 env.bak/
 venv.bak/
-lib/globals.py
+backend/globals.py
 
 
 # IDE-specific ignores (add or remove based on your IDE)
