+++ backend/globals.py
