+++ backend/window/app.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, send_from_directory, request, jsonify
 import os
 import boto3
-# from lib.globals import USERNAME
+# from backend.globals import USERNAME
 import json
 
 app = Flask(__name__)
