+++ backend/window/static/index.js
