+++ frontend/package.json
@@ -5,17 +5,20 @@
   "main": "public/electron.js",
   "scripts": {
     "start": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
     "dev": "concurrently \"vite\" \"electron .\"",
+    "build": "vite build",
     "electron:dev": "vite build && electron .",
     "electron:build": "vite build && electron-builder"
   },
   "dependencies": {
+    "@electron/remote": "^2.1.2",
+    "@tailwindcss/postcss": "^4.0.9",
+    "axios": "^1.8.1",
     "electron-is-dev": "^2.0.0",
     "lucide-react": "^0.303.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^7.2.0"
   },
   "devDependencies": {
     "@tailwindcss/vite": "^4.0.5",
@@ -24,10 +27,11 @@
     "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.20",
     "concurrently": "^9.1.2",
-    "electron": "^28.1.0",
+    "electron": "^28.3.3",
     "electron-builder": "^25.1.8",
     "electron-reloader": "^1.2.3",
     "tailwindcss": "^4.0.5",
-    "vite": "^5.0.8"
+    "vite": "^5.0.8",
+    "wait-on": "^8.0.2"
   }
 }
